fix(eq-dsp): mechanical lints to zero; 1,461 judgment lints remain for review
eq-dsp mechanical 1,051 -> 0. Verified: `cargo check -p eq-dsp --all-targets`
clean, `cargo nextest -p eq-dsp` 255/256 with the one pre-existing failure
(all_filters_all_slopes_scan, missing ref-data dir) unchanged.

This is the first run of the hardened workflow — the previous four invoked it
by name and silently got the pre-edit script (see the amended commit before
this one). Invoked by scriptPath, 24 haiku agents cleared 874 violations in
15 minutes with no build breakage and no API refactors, against 188 fixes and
three broken builds from the old loop.

The last 177 were `unreadable_literal` in the s9 coefficient tables, which no
agent should ever have been spending a turn on. They are not machine-applicable
in the sense `cargo clippy --fix` uses: the suggestion hangs off a CHILD span
marked MaybeIncorrect, so --fix skips it silently. Applied directly from
clippy's own JSON by byte span, guarded so an edit is rejected unless the
replacement is the identical numeral with underscores added — a coefficient
table is exactly where a "helpful" digit change would never be noticed.

What is left in this crate is 1,461 judgment lints, deliberately untouched:
suboptimal_flops 368, indexing_slicing 326, as_conversions 225,
arithmetic_side_effects 119, cast_* 247, float_cmp 61. Every one sits in
per-sample filter math where the correct rewrite depends on whether the code
runs on an audio callback. Those are for a human, not a fixer agent.

+++ features/fx/eq/eq-dsp/src/chain.rs
@@ -43,11 +43,13 @@ impl EqChain {
     }
 
     /// Get an immutable reference to a band by index.
+    #[must_use]
     pub fn band(&self, index: usize) -> Option<&Band> {
         self.bands.get(index)
     }
 
     /// Return the number of bands in the chain.
+    #[must_use]
     pub const fn num_bands(&self) -> usize {
         self.bands.len()
     }
@@ -79,6 +81,7 @@ impl EqChain {
     /// Each sample passes through all bands in series (left then right).
     /// True when at least one band would touch the signal.
     /// The whole chain's magnitude response at `hz`, in dB.
+    #[must_use]
     pub fn magnitude_db(&self, hz: f64, sample_rate: f64) -> f64 {
         self.bands
             .iter()
@@ -86,6 +89,7 @@ impl EqChain {
             .sum()
     }
 
+    #[must_use]
     pub fn has_active_bands(&self) -> bool {
         self.bands.iter().any(|b| !b.is_idle())
     }
@@ -174,7 +178,7 @@ mod placement_tests {
         let n = 4800;
         let (mut l, mut r) = (vec![0.0f64; n], vec![0.0f64; n]);
         for i in 0..n {
-            let x = (core::f64::consts::TAU * 100.0 * i as f64 / sr).sin();
+            let x = (core::f64::consts::TAU * 100.0 * (i as i32 as f64) / sr).sin();
             l[i] = x;
             r[i] = x;
         }
